Create utils/html.py
