+++ wagtail/utils/text.py
@@ -5,6 +5,19 @@ from anyascii import anyascii
 from django.utils.encoding import force_str
 from django.utils.text import slugify
 
+from wagtail.utils.coreutils import SCRIPT_RE
+
+SCRIPT_RE = re.compile(r"<(-*)/script>")
+
+
+def escape_script(text):
+    """
+    Escape `</script>` tags in 'text' so that it can be placed within a `<script>` block without
+    accidentally closing it. A '-' character will be inserted for each time it is escaped:
+    `<-/script>`, `<--/script>` etc.
+    """
+    return SCRIPT_RE.sub(r"<-\1/script>", text)
+
 
 def safe_snake_case(value):
     """
